requirements.txt
Django==4.0.5
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt=5.2.0